requirements.txt
fastapi[standard]==0.115.6
uvicorn==0.32.1
fastapi-keycloak-auth==1.0.0